requirements.txt
streamlit
llama-index
llama-index-core
llama-index-llms-groq
llama-index-embeddings-huggingface
python-dotenv